Java interop fix: replaced package objects with classes on the same package level
