+++ engine/api/src/main/scala/pl/touk/nussknacker/engine/api/deployment/DeploymentId.scala
@@ -1,0 +1,5 @@
+package pl.touk.nussknacker.engine.api.deployment
+
+import io.circe.generic.JsonCodec
+
+@JsonCodec case class DeploymentId(value: String) extends AnyVal
