+++ ui/restmodel/src/main/scala/pl/touk/nussknacker/restmodel/process/ProcessId.scala
@@ -1,0 +1,9 @@
+package pl.touk.nussknacker.restmodel.process
+
+import pl.touk.nussknacker.engine.api.process.ProcessName
+
+final case class ProcessId(value: Long) extends AnyVal
+
+final case class ProcessIdWithName(id: ProcessId, name: ProcessName)
+
+final case class ProcessIdWithNameAndCategory(id: ProcessId, name: ProcessName, category: String)
