--- ui/restmodel/src/main/scala/pl/touk/nussknacker/restmodel/process/package.scala
@@ -1,12 +1,0 @@
-package pl.touk.nussknacker.restmodel
-
-import pl.touk.nussknacker.engine.api.process.ProcessName
-
-package object process {
-  final case class ProcessId(value: Long) extends AnyVal
-
-  final case class ProcessIdWithName(id: ProcessId, name: ProcessName)
-
-  final case class ProcessIdWithNameAndCategory(id: ProcessId, name: ProcessName, category: String)
-
-}
